requirements.txt
aiofiles==0.8.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
altgraph==0.17.3
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncpraw==7.7.0
asyncprawcore==2.3.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
Cython==0.29.34
discord==2.2.2
discord.py==2.2.2
Django==4.1.3
dnspython==2.3.0
docopt==0.6.2
frozenlist==1.3.3
idna==3.4
image==1.5.33
macholib==1.16.2
multidict==6.0.4
Pillow==9.3.0
pipreqs==0.4.11
praw==7.7.0
prawcore==2.3.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
update-checker==0.18.0
urllib3==1.26.15
websocket-client==1.5.1
yarg==0.1.9
yarl==1.8.2
